Chain_with_LangGraph/requirement.txt
langchain
langgraph
langchain-core
langchain-community
arxiv
wikipedia
langchain-groq

streamlit
langchain
langchain-community
langchain-groq
langgraph
tavily-python
python-dotenv
ipython
pillow  # For Image display in IPython